requirements.txt
accelerate
modelscope
pytorch-lightning
imageio 
torchvision
bitsandbytes
opencv-python
open_clip_torch 
av
omegaconf
compel
diffusers==0.27.2
safetensors
decord
moviepy
pytorchvideo
vidaug
scikit-image
transformers==4.34.0
git+https://github.com/openai/CLIP.git
urllib3==1.26.7
requests==2.28.2
huggingface_hub==0.23.0
einops
